+++ src/main/java/org/kumoricon/registration/print/formatter/BadgePrintFormatter.java
@@ -1,0 +1,8 @@
+package org.kumoricon.registration.print.formatter;
+
+
+import java.io.InputStream;
+
+public interface BadgePrintFormatter {
+    InputStream getStream();
+}