+++ src/main/java/org/kumoricon/registration/utility/PrinterController.java
@@ -2,6 +2,10 @@ package org.kumoricon.registration.utility;
 
 
 import org.kumoricon.registration.controlleradvice.CookieControllerAdvice;
+import org.kumoricon.registration.controlleradvice.PrinterSettings;
+import org.kumoricon.registration.model.badge.BadgeService;
+import org.kumoricon.registration.print.BadgePrintService;
+import org.kumoricon.registration.print.PrinterInfoService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
@@ -11,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import javax.print.PrintException;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletResponse;
 import java.io.UnsupportedEncodingException;
@@ -19,38 +24,58 @@ import java.nio.charset.StandardCharsets;
 
 @Controller
 public class PrinterController {
-    private final PrinterService printerService;
+    private final PrinterInfoService printerInfoService;
+    private final BadgePrintService badgePrintService;
     private static final Logger log = LoggerFactory.getLogger(PrinterController.class);
 
-    public PrinterController(PrinterService printerService) {
-        this.printerService = printerService;
+    public PrinterController(PrinterInfoService printerInfoService, BadgePrintService badgePrintService) {
+        this.printerInfoService = printerInfoService;
+        this.badgePrintService = badgePrintService;
     }
 
     @RequestMapping(value = "/utility/printer")
     public String printer(Model model,
-                          @CookieValue(value = CookieControllerAdvice.PRINTER_COOKIE_NAME, required = false) String printerName) {
-        model.addAttribute("printer", printerName);
-        model.addAttribute("availablePrinters", printerService.getPrinterNames());
+                          @CookieValue(value = CookieControllerAdvice.PRINTER_COOKIE_NAME, required = false) String printerCookie) {
+        PrinterSettings settings = PrinterSettings.fromCookieValue(printerCookie);
+
+        model.addAttribute("printer", settings.getPrinterName());
+        model.addAttribute("xOffset", settings.getxOffset());
+        model.addAttribute("yOffset", settings.getyOffset());
+        model.addAttribute("availablePrinters", printerInfoService.getPrinterNames());
         return "utility/printer";
     }
 
     @RequestMapping(value = "/utility/printer", method = RequestMethod.POST)
     public String setPrinter(Model model,
                              @RequestParam String printer,
+                             @RequestParam(required = false, defaultValue = "0") Integer xOffset,
+                             @RequestParam(required = false, defaultValue = "0") Integer yOffset,
                              @RequestParam String action,
                              HttpServletResponse response) {
         String newPrinter = printer.trim();
         String urlSafePrinterName = getUrlSafePrinterName(newPrinter);
+        PrinterSettings settings = new PrinterSettings(newPrinter, xOffset, yOffset);
 
         if (action.equals("Test")) {
             log.info("Printed test badge to " + newPrinter);
-            // TODO: send test badge to selected printer
-            return "redirect:/utility/printer?msg=Printed+Test+Badge+to+" + urlSafePrinterName;
+            model.addAttribute("printer", printer);
+            model.addAttribute("xOffset", xOffset);
+            model.addAttribute("yOffset", yOffset);
+            model.addAttribute("availablePrinters", printerInfoService.getPrinterNames());
+            try {
+                badgePrintService.printTestBadge(settings);
+                model.addAttribute("msg", "Printed test badge to " + newPrinter);
+                return "utility/printer";
+            } catch (PrintException ex) {
+                log.error("Error printing to {}", newPrinter, ex);
+                model.addAttribute("err", ex.getMessage());
+                return "utilty/printer";
+            }
         } else if (action.equals("Save") || action.equals("Select")) {
-            log.info("Setting printer to {}", newPrinter);
+            log.info("Setting printer to {}", settings);
             try {
                 model.addAttribute("printer", newPrinter);
-                Cookie cookie = new Cookie(CookieControllerAdvice.PRINTER_COOKIE_NAME, newPrinter);
+                Cookie cookie = new Cookie(CookieControllerAdvice.PRINTER_COOKIE_NAME, settings.asCookieValue());
                 cookie.setPath("/");
                 cookie.setHttpOnly(true);
                 response.addCookie(cookie);